Gloo-only CPU-based monitored barrier
Pull Request resolved: https://github.com/pytorch/pytorch/pull/53773

Closes https://github.com/pytorch/pytorch/issues/52876

Implements a barrier by doing send/recv to rank 0, and rank 0 waits for these requests and on timeout, throws an exception indicating which rank did not join in the given timeout.

This barrier is only intended for CPU use cases and built into process group gloo, and will be used for debugging synchronization/hang issues.

Differential Revision: [D26921357](https://our.internmc.facebook.com/intern/diff/D26921357/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D26921357/)!
ghstack-source-id: 124687596